Checkbox, and move options rendering to separate component
