+++ client/index.ts
@@ -1,9 +1,9 @@
 // deno bundle --no-check --import-map import-map.json client/index.ts | terser > index.bundle.min.js
 
-import { createElement as e, ReactNode } from './lib/react.ts'
+import { createElement as e } from './lib/react.ts'
 import { render } from './lib/react-dom.ts'
 import { Offlineifier } from './components/Offlineifier.ts'
-import { Options } from './components/Options.ts'
+import { OptionsManager } from './components/OptionsManager.ts'
 
 declare global {
   interface Window {
@@ -16,7 +16,7 @@ const App = () => {
     'div',
     null,
     e(Offlineifier, { offline: !!window.offline }),
-    e(Options)
+    e(OptionsManager)
   )
 }
